+++ API/DTO/ProductRecord/ProductRecordResponse.cs
@@ -5,7 +5,7 @@ public class ProductRecordResponse
     public int Id { get; set; }
     public string Name { get; set; }
     public double Grams { get; set; }
-    public double Calories { get; set; }
+    public int Calories { get; set; }
     public double Protein { get; set; }
     public double Fat { get; set; }
     public double Carbohydrates { get; set; }
